COR-340 Fix test structure
