+++ web/app/client/.gitignore
@@ -1,2 +1,2 @@
-lib/
-node_modules/
+/lib
+/node_modules
