+++ web/app/client/src/lib/utils/index.ts
@@ -1,0 +1,1 @@
+export {getFieldKind} from "./getFieldKind";
