--- web/app/frontend/src/components/form/Switch.tsx
@@ -1,35 +1,0 @@
-import {View} from "react-native";
-import React from "react";
-import {Switch as SwitchRNP, Text} from "react-native-paper";
-import {darkTheme} from "../../constants/theme";
-import {InputProps} from "./FormControl";
-
-const Switch: React.FC<InputProps> = (
-    {
-        fieldDefinition,
-        style,
-        value,
-        onChange,
-    }) => {
-
-    return (
-        <View style={style}>
-            {fieldDefinition.label &&(
-                <Text theme={darkTheme}>
-                    {fieldDefinition.label[0].value}
-                </Text>
-            )}
-            {fieldDefinition.description &&
-            <Text theme={darkTheme} style={{fontSize: 10}}>
-                {fieldDefinition.description[0].value}
-            </Text>
-            }
-            <SwitchRNP
-                value={value}
-                onValueChange={onChange}
-            />
-        </View>
-    );
-};
-
-export default Switch;
